ARM: at91: nand: set default ecc_mode as NAND_ECC_SOFT.
since in the commit:
  403638ea9274b26f7eb82e3931813dbf589ec70f
  (mtd: at91: extract hw ecc initialization to one function)

the code in atmel_nand.c that will set default ECC as NAND_ECC_SOFT was removed.
instead, it starts to use ecc_mode to select ecc.

It impacts all atmel_nand user. So in this commit we set the default ecc mode
as NAND_ECC_SOFT in the board file.

Signed-off-by: Josh Wu <josh.wu@atmel.com>